GUIOMETRY opening... COMPLETE
That's right!

You can now save and reopen GUIOMETRY files, the only thing is the GUI
handling of file changes means that right now you are limited to using
'Save As...'

But, right now this is sweet success, there may still be a few bugs to
iron out since I haven't got complete test data... but it works for me
so far, more updates coming shortly.